♻️ [Refactor]: auth/store/auth.store.ts 제거
auth/model/auth.store.ts 사용
