+++ src/shared/utils/logout.ts
@@ -1,4 +1,4 @@
-import { useAuthStore, useTokenStore } from "@/entities/auth/store/auth.store";
+import { useAuthStore, useTokenStore } from "@/entities/auth/model/auth.store";
 import URL from "@/shared/constants/URL";
 
 export const logout = async () => {
